chore: remove unused `src/workspace-symbol.ts` file & imports
- now using `src/references.ts` with `getReferences()`
